Separate duplicate code to functions
